# src/lib/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Prophezeiung API",
    "version": "1.0.0",
    "description": "API for the Prophezeiung prediction platform"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication using JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [true]
          }
        },
        "required": ["success"]
      },
...
      "Round": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "submissionDeadline": {
            "type": "string",
            "format": "date-time"
          },
          "ratingDeadline": {
            "type": "string",
            "format": "date-time"
          },
          "fulfillmentDate": {
            "type": "string",
            "format": "date-time"
          },
          "resultsPublishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "prophecies": {
                "type": "number"
              }
            },
            "required": ["prophecies"]
          }
        },
        "required": [
          "id",
          "title",
          "submissionDeadline",
          "ratingDeadline",
          "fulfillmentDate",
          "resultsPublishedAt",
          "createdAt"
        ]
      },
      "RoundsListResponse": {
        "type": "object",
        "properties": {
          "rounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Round"
            }
          }
        },
        "required": ["rounds"]
      },
      "RoundDetailResponse": {
        "type": "object",
        "properties": {
          "round": {
            "$ref": "#/components/schemas/Round"
          }
        },
        "required": ["round"]
      },
      "CreateRoundRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "submissionDeadline": {
            "type": "string",
            "nullable": true
          },
          "ratingDeadline": {
            "type": "string",
            "nullable": true
          },
          "fulfillmentDate": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["title", "submissionDeadline", "ratingDeadline", "fulfillmentDate"]
      },
      "UpdateRoundRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "submissionDeadline": {
            "type": "string",
            "nullable": true
          },
          "ratingDeadline": {
            "type": "string",
            "nullable": true
          },
          "fulfillmentDate": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["title", "submissionDeadline", "ratingDeadline", "fulfillmentDate"]
      },
...
      "Prophecy": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "creatorId": {
            "type": "string"
          },
          "roundId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "fulfilled": {
            "type": "boolean",
            "nullable": true
          },
          "resolvedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "averageRating": {
            "type": "number",
            "nullable": true
          },
          "ratingCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "creatorId",
          "roundId",
          "createdAt",
          "fulfilled",
          "resolvedAt",
          "averageRating",
          "ratingCount"
        ]
      },
      "PropheciesListResponse": {
        "type": "object",
        "properties": {
          "prophecies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Prophecy"
            }
          }
        },
        "required": ["prophecies"]
      },
      "ProphecyDetailResponse": {
        "type": "object",
        "properties": {
          "prophecy": {
            "$ref": "#/components/schemas/Prophecy"
          }
        },
        "required": ["prophecy"]
      },
      "CreateProphecyRequest": {
        "type": "object",
        "properties": {
          "roundId": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "default": ""
          }
        },
        "required": ["roundId", "title"]
      },
...
      "ResolveRequest": {
        "type": "object",
        "properties": {
          "fulfilled": {
            "type": "boolean"
          }
        },
        "required": ["fulfilled"]
      },
...
    },
...
  },
  "paths": {
...
    "/api/rounds": {
      "get": {
        "tags": ["Rounds"],
        "summary": "Get all rounds",
        "responses": {
          "200": {
            "description": "List of rounds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Rounds"],
        "summary": "Create a new round (Admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoundRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Round created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rounds/{id}": {
      "get": {
        "tags": ["Rounds"],
        "summary": "Get round details",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Round ID"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Round details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Round not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Rounds"],
        "summary": "Update round (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Round ID"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoundRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Round updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Round not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Rounds"],
        "summary": "Delete round (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Round ID"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Round deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Round not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/prophecies": {
      "get": {
        "tags": ["Prophecies"],
        "summary": "Get prophecies with optional filters",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "roundId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["mine", "toRate"]
            },
            "required": false,
            "name": "filter",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of prophecies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropheciesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Prophecies"],
        "summary": "Create a new prophecy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProphecyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prophecy created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProphecyDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or submission deadline passed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Round not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/prophecies/{id}/resolve": {
      "post": {
        "tags": ["Prophecies"],
        "summary": "Mark prophecy as fulfilled or not (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Prophecy ID"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prophecy resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProphecyDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Rating phase not ended yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Prophecy not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}